users_analyze_api/src/requirements.txt
kafka-python==2.0.2
pydantic[email]==1.9.1
uvicorn==0.12.2
python-dotenv==1.0
fastapi==0.95.2
orjson==3.8.7
gunicorn==20.1.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
httptools==0.5.0